+++ src/components/AppHeader/AppHeader.tsx
@@ -7,7 +7,7 @@ import {
     ActionIcon,
     NumberInput,
 } from "@mantine/core";
-import { Hymn } from "../utils";
+import { Hymn } from "../../utils";
 import {
     IconChevronLeft,
     IconChevronRight,
@@ -15,7 +15,7 @@ import {
 } from "@tabler/icons-react";
 
 import classes from "./AppHeader.module.css";
-import { HymnMap } from "../PreviewContainer";
+import { HymnMap } from "../../PreviewContainer";
 
 export function AppHeader({
     drawerOpened,
