+++ src/components/HymnPreview/HymnPreview.tsx
@@ -1,9 +1,12 @@
-import { Alert, Center, Code, TypographyStylesProvider } from "@mantine/core";
+import { Center, TypographyStylesProvider } from "@mantine/core";
 import ReactMarkdown from "react-markdown";
 
 import classes from "./HymnPreview.module.css";
-import { Hymn } from "../utils";
-import { IconInfoCircle } from "@tabler/icons-react";
+import { Hymn } from "../../utils";
+import {
+    InvalidHymnMessage,
+    NoHymnMessage,
+} from "../ErrorMessages/ErrorMessages";
 
 export default function HymnPreview({ selectedItem }: { selectedItem?: Hymn }) {
     return (
@@ -26,32 +29,5 @@ export default function HymnPreview({ selectedItem }: { selectedItem?: Hymn }) {
                 <NoHymnMessage />
             )}
         </Center>
-    );
-}
-
-export function InvalidHymnMessage({ selectedHymn }: { selectedHymn: Hymn }) {
-    return (
-        <Alert
-            variant="light"
-            color="red"
-            title="Invalid hymn"
-            icon={<IconInfoCircle />}
-        >
-            This hymn (number {selectedHymn.number}) does not have a valid{" "}
-            <Code>markdown</Code> or <Code>content</Code> key
-        </Alert>
-    );
-}
-
-export function NoHymnMessage() {
-    return (
-        <Alert
-            variant="light"
-            color="gray"
-            title="No hymn selected"
-            icon={<IconInfoCircle />}
-        >
-            There is no hymn with the selected number
-        </Alert>
     );
 }
